+++ src/main/java/com/example/backend/todo/repository/TodoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface TodoRepository extends JpaRepository<Todo, Integer> {
+public interface TodoRepository extends JpaRepository<Todo, Long> {
 
     @Query("select t from Todo t where (t.todoDate>=CURRENT_DATE and t.state=true) or (t.state=false)")
     Page<Todo> findAllTodo(Pageable pageable);
